+++ Assets/Dossiers Personnels/Sarah/Script/SlimeStats_SO.cs
@@ -6,7 +6,7 @@ public class SlimeStats_SO : ScriptableObject
     [field: SerializeField]
     public float chargeSpeed { get; private set; }
     [field: SerializeField]
-    public float health { get; private set; }
+    public int health { get; private set; }
     [field: SerializeField]
     public int power { get; private set; }
     [field: SerializeField]
